pyproject.toml
[tool.poetry]
name = "projeto saja runner"
version = "0.1.0"
description = ""
authors = ["Alisson Araujo de Sousa <alissonaraujo20042018@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
streamlit = "^1.41.1"
streamlit-modal = "^0.1.2"
phonenumbers = "^8.13.53"
matplotlib = "^3.10.0"
plotly-express = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
